Collapse spring configuration into the SpringTo plan as an optional property.
Summary: This removes the ability to change the configuration of a spring without also changing its destination. We feel that such a situation may be strictly hypothetical, so in the interest of simplifying the API we now have one plan that requires a destination and an optional configuration.

Reviewers: O4 Material Motion Apple platform reviewers, O2 Material Motion, markwei

Reviewed By: O4 Material Motion Apple platform reviewers, O2 Material Motion, markwei

Tags: #material_motion

Differential Revision: http://codereview.cc/D1701