Add PauseSpring.
Summary:
The purpose of this plan is to allow a director to keep springs from taking effect while gesture recognizers are active.

Spec: https://material-motion.github.io/material-motion/starmap/specifications/plans/PauseSpring

This plan could be generalized in the future to accept some form of "is-active" stream if we wanted springs to be dependent on other types of stream events as well. As we encounter real situations that need this we'll re-evaluate the API and generalize it accordingly.

The internal workings of the pop performer have been adjusted to allow springs to be removed and re-added in reaction to gesture events. We now keep track of a small amount of in-flight state such as the in-flight POP animation, whether any gesture recognizers that affect its behavior are active, and the uuid key we use to refer to the animation in POP.

Reviewers: markwei, O4 Material Motion Apple platform reviewers, O2 Material Motion

Reviewed By: markwei, O4 Material Motion Apple platform reviewers, O2 Material Motion

Subscribers: markwei

Tags: #material_motion

Maniphest Tasks: T84

Differential Revision: http://codereview.cc/D1908