Move PauseSpring and SpringTo APIs to a string-based property API.
Summary:
This is a breaking change.

Closes https://github.com/material-motion/pop-swift/issues/24

### Why this change is happening

This change brings SpringTo and PauseSpring's property API in alignment with the Core Animation Tween property API and encourages a more consistent way to use Material Motion plans.

### How this change was made

POP has out-of-the-box support for a subset of key paths for certain object types, so we map the provided property key path to these supported types. We still support the same number of properties, but we no longer clearly communicate which properties are supported and now throw a runtime assertion if an unsupported key path is provided. https://github.com/material-motion/pop-swift/issues/25 tracks the addition of a fall-through case for other key paths.

Reviewers: O2 Material Motion, O4 Material Motion Apple platform reviewers, markwei

Reviewed By: O2 Material Motion, O4 Material Motion Apple platform reviewers, markwei

Tags: #material_motion

Maniphest Tasks: T86

Differential Revision: http://codereview.cc/D1960